Long to int

+++ app/src/main/java/fr/neamar/lolgamedata/pojo/Summoner.java
@@ -6,12 +6,12 @@ import org.json.JSONObject;
 import java.io.Serializable;
 
 public class Summoner implements Serializable {
-    public long id;
+    public int id;
     public String name;
     public int level;
 
     public Summoner(JSONObject summoner) throws JSONException {
-        id = summoner.getLong("id");
+        id = summoner.getInt("id");
         name = summoner.getString("name");
         level = summoner.getInt("level");
     }
